Add lines nodes & json types & nodes
We change the standard input to return the whole input as text so that
it is available for parsing non-single line entries and then we
introduce a json parse node to take advantage of that. Then we make a
json keys node to have something to mess around with the json data.