Add multi-input node example
We implement 'json-object' a node that takes two inputs. A list of
string keys and a list of string values and creates a json object by
pairing them up.

We also implement a 'json-stringify' node so that we convert the result
back to text to print it smoothly to standard out.