Super hacky working node connections
I haven't dealt with visualising the connections at all but we can tell
enough from the various events that are going on to connect the nodes
together based on a ctrl-click & drag.

We then assume that the last node created is the node to pull on and so
do that in the 'exit' bit of the code that responds to the 'q' key.
Incredibly hacky but it actually works which is pretty crazy exciting
to be honest.