Draws line when connecting nodes
All kinds of trouble here as the line becomes the 'widget_under_cursor'
so we can no longer rely on that method and instead have to walk the
widget graph and do bounding box checks on all the nodes we encounter.

Lots of trial & error and it kind of works except that everything is
drawn in the wrong space so the final line showing the connection is way
off.