Change to relative coordinates & layout
We had an issue before where the main widgets think about centered axis
and +/- values for position where as the glium mouse events were coming
in with top-left-corner axis and only positive values so we were having
to convert between them.

Here we switch to conrod events so that the mouse has the same relative
positioning as everything else and we can use the values more easily
without conversion.