Add working undo/redo mechanism
Probably not perfect as I am still finding my way but it seems
promising. I've ended up passing in the 'state' or 'params' to be edited
into the undo/redo rather than having the nodes own a reference to the
state as that was looking pretty hairy with lifetimes and things.

Pretty proud that this works. I'm pursuing vim bindings where possible
so we have 'u' and 'ctrl-r' for undo & redo at the moment. I also have
'escape' to close the node creation box without exiting the program.