Split files up a bit
Mostly trying to move the commands stuff out of the gui file. Ended up
moving the params definition as well. Not really sure about it but it is
a little more manageable. Hoping to do a clean refactor in the future
when I have a better idea of how to structure & name rust
projects/modules.