Add 'background' widget
I want to be able to detect clicks on the background of the node graph
to figure out when we're deselecting nodes. It seems that the default
'canvas' widget doesn't allow us to detect clicks so we make a
Background widget that wraps a single canvas and exposes any clicks that
happen on it.