Hack together working parameter for 'string-contains'
We focus on that node and the single value that it expects but try to
implement it in a relatively generic way as far as the node goes.

I'm somewhat confused by how to handle the conrod widget ids for the
various bits on interface that I'm going to generate so that side is
very hacky indeed and will need a proper solution before too long.