Allow 'i' to insert before node
And if there is already a node connected to the current node then we
break that connection and connect up the new node to both the input node
and the target node. So that 1 -> 2 becomes 1 -> 3 -> 2.

We have to introduce a 'disconnect' command so that it is undoable. We
also switch the connection list to a hashmap keyed on the (from, to) in
order to make it easier to add & remove connections.

I struggled to get the connections to insert into the command in the
disconnect 'undo' so I had to create a new connection for it but it is
pretty much just plain old data so that isn't really an issue.