Rewrite 'save' as a Command
And push on to undo stack.

And update the Command trait to have execute, redo & undo which is based
on a vague memory from working with commands in the Maya API. There are
sometimes things that you want to store once in the execute and not to
redo everytime you the command is redone.