Create gui nodes & connections on load
In order to achieve this we need to save the node labels into the file
and save the positions more convincingly as floats rather than without
any precision as then they are intepreted as ints in the file and the
yaml reading fails if we expect them to be a float.

We also move the reading code out into a separate function as it is a
good thing and to resolve a double mutable ownership thing.